langchain-course/intro-to-vector-dbs/pyproject.toml
[project]
name = "intro-to-vector-dbs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "faiss-cpu>=1.11.0",
    "isort>=6.0.1",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-community>=0.3.23",
    "langchain-huggingface>=0.2.0",
    "langchain-ollama>=0.3.2",
    "langchain-pinecone>=0.2.6",
    "langchain-voyageai>=0.1.4",
    "langchainhub>=0.1.21",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
